agent/requirements.txt
anthropic
openai
yfinance
pandas
matplotlib
plotly
pdfplumber
python-docx
python-dotenv
kafka-python
redis
psycopg2-binary
pgvector
voyageai
requests
aiofiles>=23.0
